Add timing file
This timing file was entered into the amplitude modulated regression. The first number is the time of the onset of the word, the second number is the lexical frequency of that word, and the third number is the phonological neighborhood density of that word. Both lexical properties were log transformed and z-scored